remove circular reference
